# backend/Documentation/Swagger.json
{
    "openapi": "3.0.5",
    "info": {
        "version": "1.0.0",
        "title": "My User Project CRUD",
        "description": "My User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "http://192.168.29.128:4000/new-docs/",
    "basePath": "/api/",
    "tags": [
        {
            "name": "Users",
            "description": "API for users in the system"
        },
        {
            "name": "Employee",
            "description": "API for employees in the system"
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "server":[
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "Add new User",
                "parameters": [
                    {
                        "name": "",
                        "in": "body",
                        "description": "User that we want to create",
                        "schema": {
                            "$ref": "#/definitions/User/UserReg"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "New user is created"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "To Loging in user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "",
                        "in": "body",
                        "description": "User that we want to create",
                        "schema": {
                            "$ref": "#/definitions/User/UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verifying User",
                        "schema": {
                            "$ref": "#/definitions/User/UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/User/UserLoginResponseError"
                        }
                    },
                    "417": {
                        "description": "Exception Failed",
                        "schema": {
                            "$ref": "#/definitions/User/UserLoginResponseError"
                        }
                    }
                }
            }
        },
        "/api/refresh": {
            "post": {
                "tags": [
                    "Users"
                ],
                "description": "To Refresh the token",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "",
                        "in": "body",
                        "description": "JWT token Refresher",
                        "schema": {
                            "$ref": "#/definitions/User/UserRefreshToken"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Token refreshed",
                        "schema": {
                            "$ref": "#/definitions/User/UserLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/getAdminUser": {
            "get": {
                "tags": [
                    "Users"
                ],
                "description": "To Get the all Admin Users",
                "security":{
                    "Bearer":["Users"]
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User/Users"
                        }
                    }
                }
            }
        },
        "/api/randomPass": {
            "get": {
                "tags": [
                    "Users"
                ],
                "security":{
                    "Bearer":["Users"]
                },
                "description": "To generate random passwords",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User/UserRandomPassword"
                        }
                    }
                }
            }
        },
        "/api/employee": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "description": "To Get All the Employees",
                "security":{
                    "Bearer":["Users"]
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeGet"
                        }
                    }
                }
            }
        },
        "/api/employee/show": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "description": "To Get single Employees",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "",
                        "in": "body",
                        "produces": [
                            "application/json"
                        ],
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeShow"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "the employee",
                        "schema": {
                            "items": {
                                "$ref": "#/definitions/Employee/EmployeeGet"
                            }
                        }
                    }
                }
            }
        },
        "/api/employee/store": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "description": "To Add a new Eployee",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeStore"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "201": {
                        "description": "Added succesfully"
                    },
                    "406": {
                        "description": "Wrong Parameters"
                    }
                }
            }
        },
        "/api/employee/update": {
            "put": {
                "tags": [
                    "Employee"
                ],
                "description": "To Update an Employee",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated!!!"
                    },
                    "304": {
                        "description": "Not Acceptable"
                    }
                }
            }
        },
        "/api/employee/delete": {
            "delete": {
                "tags": [
                    "Employee"
                ],
                "description": "To Delete an employee",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeShow"
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "the record has been deleted successfully"
                    }
                }
            }
        },
        "/api/employee/customPagination": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "description": "To Get All the Employees Pagination wise",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "integer",
                        "required": true
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "type": "integer",
                        "required": true
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Employee/CustomPagination"
                        }
                    }
                }
            }
        },
        "/api/employee/sortRecs": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "description": "To Sort the records",
                "security":{
                    "Bearer":["Users"]
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeGet"
                        }
                    }
                }
            }
        },
        "/api/employee/sortWithPagination": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "description": "To sort records with pagination",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeGet"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeGet"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeGet"
                        }
                    },
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "enum": [
                            "ascending",
                            "descending"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Employee/CustomPagination"
                        }
                    }
                }
            }
        },
        "/api/employee/searchRec": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "description": "To Find a particular employee",
                "security":{
                    "Bearer":["Users"]
                },
                "parameters": [
                    {
                        "name": "searchBy",
                        "in": "query",
                        "required": true
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": true
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Found a Record",
                        "schema": {
                            "$ref": "#/definitions/Employee/EmployeeGet"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "UserReg": {
                "required": [
                    "name",
                    "email",
                    "mobile",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserLogin": {
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "UserLoginResponseError": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                }
            },
            "UserLoginResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
            "UserRefreshToken": {
                "properties": {
                    "refreshToken": {
                        "type": "string",
                        "description": "Ener your token here"
                    }
                }
            },
            "Users": {
                "required": [
                    "name",
                    "email",
                    "mobile",
                    "password"
                ],
                "properties": {
                    "_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    }
                }
            },
            "UserRandomPassword": {
                "properties": {
                    "result": {
                        "type": "string"
                    }
                }
            }
        },
        "Employee": {
            "EmployeeGet": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "password"
                ],
                "properties": {
                    "_id": {
                        "type": "string",
                        "uniqueItems": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "updatedAt": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string"
                    }
                }
            },
            "EmployeeShow": {
                "required": [
                    "employeeId"
                ],
                "properties": {
                    "employeeId": {
                        "type": "string"
                    }
                }
            },
            "EmployeeStore": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer"
                    },
                    "designation": {
                        "type": "string"
                    }
                }
            },
            "EmployeeUpdate": {
                "required": [
                    "employeeId",
                    "name",
                    "email",
                    "phone",
                    "password"
                ],
                "properties": {
                    "employeeId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mobile": {
                        "type": "string"
                    },
                    "age": {
                        "type": "integer"
                    },
                    "designation": {
                        "type": "string"
                    }
                }
            },
            "CustomPagination": {
                "properties": {
                    "data": {
                        "type": "array",
                        "$ref": "#/definitions/Employee/EmployeeGet"
                    },
                    "totalRec": {
                        "type": "integer"
                    },
                    "currentPage": {
                        "type": "integer"
                    },
                    "limit": {
                        "type": "integer"
                    },
                    "totalPage": {
                        "type": "integer"
                    },
                    "hasNextPage": {
                        "type": "boolean"
                    },
                    "hasPrevPage": {
                        "type": "boolean"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}